Added the files for creating packets.
Header is pretty much complete. Need to finish packet and fragment, but will require parting the config file to intake the maxPayload param. After getting maxPayload, the packet class will know how to fragment. The packet class basically will create an array of fragments, where fragments have headers with parts of the message. You initialize a packet with the message you want to send to the next node, and it can create (eventually) an array of strings that will represent packets over the wire. Still need to decide if maxPayload is the total size of the string, or the total size of the message.... I think?