Dumptable (#2)
* Slight changes

Cleaned up code a little. Added a Node object that will store all the
information of that specific Node. Added correct commands to implement
in file. Need to figure out what to store in Node objects and how to
create routing table for each Node.

* Added more code lol

Added Utility class file that will hold all of our utilities. Currently
Utility holds the methods to read in the configuration file and read in
the nodes file.

We just have to create the routing table and connect edges to the nodes
now. The hard part haha.