Lots of additions in functions.
Packet Class now fragments a passed message into an array of fragments, with properly configured headers and their section of the message
Node.rb now has two threads, one for server-like operations, and another for client-like operations. The client connection thread acts on only operation for now, EDGEB. But later, this will sent packets by their array of fragments. The c thread holds a hash of nodenames to their socket connections, to properly address sending in the future. 
The server thread is yet to be fully implemented. It needs to listen on its host port for incoming messages, and have those messages addressed somehow.
Near no changes to Utilities, Raxce, consider making the port that you parse out straight an int, so I don't have to make it a to_i in the middle of my connection stuff.