Update for EDGEB and DUMPTABLE
CommandQueue will hold all the commands that get called and be handled
accordingly. Routing table will be a Hash that points to an array that
will hold the nextHop and distance. It will be written to the file
according to the function in the Utility.rb file.

I made CommandHandler its own method to keep the setup method short and
clean. It only checks for EDGEB command currently, but it still needs to
be worked on for EDGEB to work correct entirely.