Significant Progress
Here's what hasn't changed:
Routing Table is basically the same.
The Controller was not broken, but it is VERY finicky. You MUST put it
to sleep a lot, or else it will shit itself. The controller is running
multiple nodes with like, 7 threads.

Here's the changes:
-Creation of two new messages:
LSU - Link State Update
LSUR - Link State Update Request
LSU using the params other than the command  to update the route on it
routing table. It checks the value of the sequence number of the update
to see if it should update or not
LSUR is a request that a node sends out to it's nextHop nodes. It
basically asks for their routing tables, to change it's own if need
arises.

Other things:
There are two new threads, one to handle time, another to handle the
outgoing LSURs.
open_sock is now nextHop_neighbors, with a twist: it strictly holds the
neighbors as a key to and array of the cost from the current node to the
neightbor, and a the open socket to it.
Prolly some other stuff I'm too tired to think of.

Things to be fixed:
EVERYTHING. Just kidding.
Currently the LSUR thread doesn't appear to be working correctly. My
recommendation is to test it using different nodes on CORE insterad of
the controller, but that may take an hour. Another thing you could do is
check if the LSUR's are making it to their destination nodes, and at the
destinations nodes, being properly converted into LSU's being sent back.
Also, make sure it works, or change LSUR vs LSU's names. Without
regex's, the only thing not breaking everything is putting the
.include?(LSUR) before the .include?(LSU) so there isn't crossover.

There's a lot, and I don't expect you to work on it during Thanksgiving.
Fuck these awful teachers/TAs, we'll get our Part 0's back by the time
Part 2 is due. Like jesus christ