Slight changes again
Added document to maybe help as a guideline on how to implement link
state routing.

Added INFNITY constant to indicate in the routing table whether or not a
path is possible from SRC to DST.

Server should be receiving packets that will have the routing
information that should be forwarded to other nodes. So, we'll probably
have to create more commands in commandHandler to process link state
packets and forwarding.

STATUS command updated to correctly show only the neighbors to node
(which can be figured out by current sockets that have been opened from
the SRC node).

http://www.maxburstein.com/blog/introduction-to-graph-theory-finding-shortest-path/

Was looking at this guide on how to create a graph and use Dijkstra's.
May be useful, but have to implement it without using PriorityQueue,
which is useless for us anyway.

Gonna be working another 12-hour shift so I don't know how much I'll be
able to work on this. Thanks for working so hard Kyle.