Implemented link state routing and dijkstra's
Link state packets will be created for each node's neighbors and flooded
out to the rest of the network.

$lst_received will keep track of all nodes it receives packets from to
avoid sending the same packet to the node it got it from.

Initially, node will have sequence number of 0 and the first set of link
state packets to be sent out will have sequence number 1.

After every update interval, dijkstra's will be run on the graph to get
the shortest path from the source node to all other nodes.

Link state packets will only be sent out if that packet didn't come from
itself, if it's a new sequence number, and if it hasn't already sent out
that specific packet to that node.